added vibrato experiment (don't work right now)
Apparently I use double standards for normalizing values

For inputs to generators 0 == 0, 1 == 22050; output 0 == -1, 1 == 1; for filters both I/O means 0 == -1, 1 == 1, 0.5 == 0

You can see why the output of a filter doesn't work as input to a generator; need to normalize input to generators to work with the same normalized values as everything else

Then there's the same issue repeated for filter parameters; K == 0 means 0, but IN == 0 means -1; oups

The new strategy is to use 0 == -1, 0.5 == 0, 1 == 1 OR 0 == -22050, 0.5 == 0, 1 == 22050 for absolutely everything: all inputs/outputs, all secondary parameters including but not limited to A, D, S, R, K, Low, High

Another problem: WT input is currently limited to the very nice, yet very limited set of { -1, 0, 1 }; ah, the flexibility; fix this to use +/- 127

Added new LOG statements; update this to use the masked LOGF instead. While on the subject, add a command line parameter to set the log/trace mask (that'll be a chore)

Consider this commit message a TODO list.

Oh, btw, delay now accepts IN as a parameter. It is no longer deaf mute.

----

Bottom line: absolutely all non-bool non-enum parameters to everything in the back end will use the convention that the input is between 0 and 1 and this gets linearly mapped to [0, 1] -> [-1, 1] or [0, 1] -> [-22050, 22050] where applicable; in the input file, numbers shall use the following range:
frequencies: 0 to 22050, in Hz (Low, High, inputs)
time values: 0 to 2550, in ms (A, D, R)
gain factors: 0 to 999, in thousandths of a unit (S, K)
Generator.Glide: 0 to infinity, in number of samples

Proposal: merge time values and Glide into a single concept: 0 to infinity, in number of samples; it makes math a bit harder, but it reduces complexity

alzwded out